I changed some casts from (usigned int) to (unsigned long) to remove errors on 64 bit compilers

+++ trunk/interface/ObjList.cpp
@@ -97,7 +97,7 @@ void ObjList::glCommands(bool select, bool marked, bool no_color)
 		HeeksObj* object = *It;
 		if(object->OnVisibleLayer() && object->m_visible)
 		{
-			if(select)glPushName((unsigned int)object);
+			if(select)glPushName((unsigned long)object);
 #ifdef HEEKSCAD
 			(*It)->glCommands(select, marked || wxGetApp().m_marked_list->ObjectMarked(object), no_color);
 #else
